refaster(power-assert-context-traversal): inline esnode.js
